src/regular/requirements.txt
pytelegrambotapi
aiohttp
python-dotenv
innertube